requirements.txt
Flask 
python-dotenv
pandas 
requests 
schedule 
Flask-SQLAlchemy 
psycopg2-binary
